fixed routing such that suggestions is now a menu, suggestions/newSuggestion is now where you create a new suggestion
