+++ app/(dashboard)/suggestions/[slug]/page.tsx
@@ -1,7 +1,7 @@
 // import { getSuggestion } from '@/lib/db';
 import Link from 'next/link';
 
-export default async function ArticleDetailPage({
+export default async function SuggestionDetailPage({
   params
 }: {
   params: Promise<{ slug: string }>;
